+++ Doc/spelling_wordlist.txt
@@ -1,5 +1,6 @@
 args
 async
+asyncsearch
 attr
 attrlist
 attrList
