+++ Tests/t_untested_mods.py
@@ -1,5 +1,5 @@
 # modules without any tests
-import ldap.async
+
 import ldap.controls.deref
 import ldap.controls.openldap
 import ldap.controls.ppolicy
@@ -9,6 +9,7 @@ import ldap.controls.readentry
 import ldap.controls.sessiontrack
 import ldap.controls.sss
 import ldap.controls.vlv
+import ldap.constants
 import ldap.logger
 import ldap.resiter
 import ldap.syncrepl
